frontend_requirements.txt
aiofiles==24.1.0
blinker==1.8.2
click==8.1.7
Flask==3.0.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
Hypercorn==0.17.3
hyperframe==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
priority==2.0.0
Quart==0.19.6
quart-auth==0.10.1
quart-uploads==0.0.2
quart-wtforms==1.0.1
Werkzeug==3.0.4
wsproto==1.2.0
WTForms==3.1.2
